feat(ecs): update ShootSystem to support cooldown as float and improv…

+++ ecs/components/Shoot.cpp
@@ -30,7 +30,7 @@ namespace ecs {
      * @brief Gets the shoot cooldown
      * @return The cooldown of the shoot
      */
-    int Shoot::getCooldown() const
+    float Shoot::getCooldown() const
     {
         return _cooldown;
     }
@@ -39,7 +39,7 @@ namespace ecs {
      * @brief Sets the shoot cooldown
      * @param cooldown New cooldown of the shoot
      */
-    void Shoot::setCooldown(int cooldown)
+    void Shoot::setCooldown(float cooldown)
     {
         _cooldown = cooldown;
     }
@@ -48,7 +48,7 @@ namespace ecs {
      * @brief Gets the last shot
      * @return The last shot
      */
-    int Shoot::getTimeSinceLastShot() const
+    float Shoot::getTimeSinceLastShot() const
     {
         return _timeSinceLastShot;
     }
@@ -57,7 +57,7 @@ namespace ecs {
      * @brief Sets the last shot
      * @param t New shot of the shoot
      */
-    void Shoot::setTimeSinceLastShot(int t)
+    void Shoot::setTimeSinceLastShot(float t)
     {
         _timeSinceLastShot = t;
     }
